Introduce new gamble command.
For now, this includes a simple storage mechanism, which currently is
just a json file. This file stores the users unique ID, which is the key
for each entry.

For human-readability, the username is also stored so any users can
easily see which ID is which user. The username may change if the user
namechanges.

At the moment, the only subcommand available is a coinflip command.